refactor: fix AnimationCurve folder bug
