+++ packages/core/src/animation/AnimationClipCurveBinding.ts
@@ -1,7 +1,7 @@
 import { Component } from "../Component";
 import { Entity } from "../Entity";
-import { AnimationCurve } from "./AnimationCurve";
-import { IAnimationCurveCalculator } from "./AnimationCurve/interfaces/IAnimationCurveCalculator";
+import { AnimationCurve } from "./animationCurve/AnimationCurve";
+import { IAnimationCurveCalculator } from "./animationCurve/interfaces/IAnimationCurveCalculator";
 import { AnimationCurveOwner } from "./internal/AnimationCurveOwner/AnimationCurveOwner";
 import { KeyframeValueType } from "./Keyframe";
