+++ packages/core/src/animation/animationCurve/AnimationArrayCurve.ts
