+++ packages/core/src/animation/animationCurve/AnimationBoolCurve.ts
