+++ packages/core/src/animation/animationCurve/AnimationColorCurve.ts
