+++ packages/core/src/animation/animationCurve/AnimationCurve.ts
