+++ packages/core/src/animation/animationCurve/AnimationFloatArrayCurve.ts
