+++ packages/core/src/animation/animationCurve/AnimationFloatCurve.ts
