+++ packages/core/src/animation/animationCurve/AnimationQuaternionCurve.ts
