+++ packages/core/src/animation/animationCurve/AnimationVector2Curve.ts
