+++ packages/core/src/animation/animationCurve/AnimationVector3Curve.ts
