+++ packages/core/src/animation/animationCurve/AnimationVector4Curve.ts
