+++ packages/core/src/animation/animationCurve/index.ts
