+++ packages/core/src/animation/animationCurve/interfaces/IAnimationCurveCalculator.ts
